GetMemfileOverprovisioningPercentage returns size_t

+++ ecal/core/include/ecal/ecal_config.h
@@ -99,7 +99,7 @@ namespace eCAL
     ECAL_API TLayer::eSendMode GetPublisherUdpMulticastMode         ();
 
     ECAL_API size_t            GetMemfileMinsizeBytes               ();
-    ECAL_API int               GetMemfileOverprovisioningPercentage ();
+    ECAL_API size_t            GetMemfileOverprovisioningPercentage ();
     ECAL_API int               GetMemfileAckTimeoutMs               ();
     ECAL_API bool              IsMemfileZerocopyEnabled             ();
     ECAL_API size_t            GetMemfileBufferCount                ();
